@@ -1,11 +1,8 @@
-import assert from 'power-assert';
+const assert = require('assert');
+const Payjp = require('../built');
+const config = require('./config');
 
-import Requestor from '../built/requestor';
-import Payjp from '../built';
-
-import config from './config';
-
-const payjp = new Payjp(config.auth_key, config);
+const cards = Payjp(config.apikey, config).customers.cards;
 
 describe('Cards Resource', () => {
 
@@ -13,7 +10,7 @@ describe('Cards Resource', () => {
   var _endpoint;
 
   before(() => {
-    Requestor.prototype.request = (...args) => {
+    cards.request = (...args) => {
       _method = args[0];
       _endpoint = args[1];
       return Promise.resolve();
@@ -22,12 +19,7 @@ describe('Cards Resource', () => {
 
   describe('create', () => {
     it('Sends the correct request', () => {
-      const query = {
-        number: 4242424242424242,
-        exp_month: 12,
-        exp_year: 2035,
-      };
-      return payjp.customers.cards.create('cus_id123', query).then(() => {
+      return cards.create('cus_id123', {}).then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'customers/cus_id123/cards');
       });
@@ -36,7 +28,7 @@ describe('Cards Resource', () => {
 
   describe('list', () => {
     it('Sends the correct request', () => {
-      return payjp.customers.cards.list('cus_id123').then(() => {
+      return cards.list('cus_id123').then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'customers/cus_id123/cards');
       });
@@ -45,7 +37,7 @@ describe('Cards Resource', () => {
 
   describe('retrieve', () => {
     it('Sends the correct request', () => {
-      return payjp.customers.cards.retrieve('cus_id123', 'id456').then(() => {
+      return cards.retrieve('cus_id123', 'id456').then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'customers/cus_id123/cards/id456');
       });
@@ -54,10 +46,7 @@ describe('Cards Resource', () => {
 
   describe('update', () => {
     it('Sends the correct request', () => {
-      const query = {
-        address_state: 'tokyo'
-      };
-      return payjp.customers.cards.update('cus_id123', 'id456', query).then(() => {
+      return cards.update('cus_id123', 'id456', {}).then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'customers/cus_id123/cards/id456');
       });
@@ -66,7 +55,7 @@ describe('Cards Resource', () => {
 
   describe('delete', () => {
     it('Sends the correct request', () => {
-      return payjp.customers.cards.delete('cus_id123', 'id456').then(() => {
+      return cards.delete('cus_id123', 'id456').then(() => {
         assert(_method === 'DELETE');
         assert(_endpoint === 'customers/cus_id123/cards/id456');
       });
