@@ -1,4 +1,4 @@
-export default {
-  auth_key: 'sk_test_c62fade9d045b54cd76d7036',
-  apibase: null
+module.exports = {
+  apikey: 'sk_test',
+  config: {},
 };
