@@ -1,11 +1,8 @@
-import assert from 'power-assert';
+const assert = require('assert');
+const Payjp = require('../built');
+const config = require('./config');
 
-import Requestor from '../built/requestor';
-import Payjp from '../built';
-
-import config from './config';
-
-const payjp = new Payjp(config.auth_key, config);
+const {events} = Payjp(config.apikey, config);
 
 describe('Events Resource', () => {
 
@@ -13,7 +10,7 @@ describe('Events Resource', () => {
   var _endpoint;
 
   before(() => {
-    Requestor.prototype.request = (...args) => {
+    events.request = (...args) => {
       _method = args[0];
       _endpoint = args[1];
       return Promise.resolve();
@@ -22,7 +19,7 @@ describe('Events Resource', () => {
 
   describe('list', () => {
     it('Sends the correct request', () => {
-      return payjp.events.list().then(() => {
+      return events.list().then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'events');
       });
@@ -31,7 +28,7 @@ describe('Events Resource', () => {
 
   describe('retrieve', () => {
     it('Sends the correct request', () => {
-      return payjp.events.retrieve('id123').then(() => {
+      return events.retrieve('id123').then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'events/id123');
       });
