@@ -1,11 +1,8 @@
-import assert from 'power-assert';
+const assert = require('assert');
+const Payjp = require('../built');
+const config = require('./config');
 
-import Requestor from '../built/requestor';
-import Payjp from '../built';
-
-import config from './config';
-
-const payjp = new Payjp(config.auth_key, config);
+const {subscriptions} = Payjp(config.apikey, config);
 
 describe('Subscription Resource', () => {
 
@@ -14,7 +11,7 @@ describe('Subscription Resource', () => {
   var _query;
 
   before(() => {
-    Requestor.prototype.request = (...args) => {
+    subscriptions.request = (...args) => {
       _method = args[0];
       _endpoint = args[1];
       if (Object.keys(args).length > 2) {
@@ -26,7 +23,7 @@ describe('Subscription Resource', () => {
 
   describe('list', () => {
     it('Sends the correct request', () => {
-      return payjp.subscriptions.list().then(() => {
+      return subscriptions.list().then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'subscriptions');
       });
@@ -39,7 +36,7 @@ describe('Subscription Resource', () => {
         customer: 'cus_id456',
         plan: 'pln_id789'
       };
-      return payjp.subscriptions.create(query).then(() => {
+      return subscriptions.create(query).then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'subscriptions');
       });
@@ -51,7 +48,7 @@ describe('Subscription Resource', () => {
       const query = {
         plan: 'pln_id789'
       };
-      return payjp.subscriptions.update('id123', query).then(() => {
+      return subscriptions.update('id123', query).then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'subscriptions/id123');
       });
@@ -60,7 +57,7 @@ describe('Subscription Resource', () => {
 
   describe('retrieve', () => {
     it('Sends the correct request', () => {
-      return payjp.subscriptions.retrieve('id123').then(() => {
+      return subscriptions.retrieve('id123').then(() => {
         assert(_method === 'GET');
         assert(_endpoint === 'subscriptions/id123');
       });
@@ -69,7 +66,7 @@ describe('Subscription Resource', () => {
 
   describe('pause', () => {
     it('Sends the correct request', () => {
-      return payjp.subscriptions.pause('id123').then(() => {
+      return subscriptions.pause('id123').then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'subscriptions/id123/pause');
       });
@@ -78,7 +75,7 @@ describe('Subscription Resource', () => {
 
   describe('resume', () => {
     it('Sends the correct request', () => {
-      return payjp.subscriptions.resume('id123').then(() => {
+      return subscriptions.resume('id123').then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'subscriptions/id123/resume');
       });
@@ -87,27 +84,9 @@ describe('Subscription Resource', () => {
 
   describe('cancel', () => {
     it('Sends the correct request', () => {
-      return payjp.subscriptions.cancel('id123').then(() => {
+      return subscriptions.cancel('id123').then(() => {
         assert(_method === 'POST');
         assert(_endpoint === 'subscriptions/id123/cancel');
-      });
-    });
-  });
-
-  describe('customer\'s subscription list', () => {
-    it('Sends the correct request', () => {
-      return payjp.customers.subscriptions.list('cus_id456').then(() => {
-        assert(_method === 'GET');
-        assert(_endpoint === 'customers/cus_id456/subscriptions');
-      });
-    });
-  });
-
-  describe('customer\'s subscription retrieve', () => {
-    it('Sends the correct request', () => {
-      return payjp.customers.subscriptions.retrieve('cus_id456', 'id123').then(() => {
-        assert(_method === 'GET');
-        assert(_endpoint === 'customers/cus_id456/subscriptions/id123');
       });
     });
   });
@@ -117,7 +96,7 @@ describe('Subscription Resource', () => {
       const query = {
         prorate: true
       };
-      return payjp.subscriptions.delete('id123', query).then(() => {
+      return subscriptions.delete('id123', query).then(() => {
         assert(_method === 'DELETE');
         assert(_endpoint === 'subscriptions/id123');
         assert(_query === query);
